chore: add optional keyRecord property to TCKey. And little polish for ProfileService.
